+++ docs/release-notes.md
@@ -14,6 +14,7 @@ These changes have not been released yet.
   * Added `IContentPack.HasFile` method.
   * Added `Context.IsGameLaunched` field.
   * Mods are now loaded much earlier in the game launch. This lets mods intercept any content asset, but the game is not fully initialised when `Entry` is called (use the `GameLaunched` event if you need to run code when the game is initialised).
+  * When a mod is incompatible, the trace logs now list all detected issues instead of the first one.
   * Dropped support for all deprecated APIs.
   * Updated to Json.NET 12.0.1.